Rename Moon component to Thing3D
we're more than just the moon now. we're Mercury
